backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
pymupdf
langchain
chromadb
sentence-transformers
langchain-community
 
